refactor(pfb): extract layout mutations into a useLayoutMutations composable
